# `bazel run` and `cargo run` are incompatible