--- a/src/main.rs
+++ b/src/main.rs
@@ -1,6 +1,7 @@
-mod app;
+// mod app;
 
-use app::app::app;
+// use app::app::app;
+use app::app;
 use axum::extract::WebSocketUpgrade;
 use axum::response::Html;
 use axum::routing::get;